Adição da tela de apresentação do exemplo.

+++ engine/include/mousebuttonevent.h
@@ -16,19 +16,19 @@ public:
     typedef enum {PRESSED, RELEASED} State;
     typedef enum { LEFT, RIGHT, MIDDLE } Button;
 
-    MouseButtonEvent(State state, Button button, int x, int y);
+    MouseButtonEvent(State state, Button button, double x, double y);
 
     State state() const;
     Button button() const;
-    int x() const;
-    int y() const;
+    double x() const;
+    double y() const;
 
     static MouseButtonEvent from_SDL(const SDL_Event& event);
 
 private:
     State m_state;
     Button m_button;
-    int m_x, m_y;
+    double m_x, m_y;
 };
 
 #endif
